{topost} serial: 8250_hisi_lpc: fix panic without lpc node
When 8250_hisi_lpc enabled, but without relevant node in DT or ACPI,
the iobase in struct uart_port is not initialized, so io_serial_out
will fail and lead to panic.

Fix this by check whether arm64_isa_pio is set or not, and only enable
HISILICON_LPC in arm64.

Signed-off-by: Kefeng Wang <wangkefeng.wang@huawei.com>