{noup} dts: hisi: fix the bug of serdes base address
Hilink4 connected to port0~3 of HNS while Hilink3 connected to port4~7.
And serdes_ctrl0 is base address for Hilink3 while serdes_ctrl1 is base
address for Hilink4. So the reference of port0~3 should be serdes_ctrl1
while port4~7 should be serdes_ctrl0.
And in hip06 the Hlink3 and Hlink4 have the same base address with an
access sel cfg register to distinguish them.

Signed-off-by: Kejian Yan <yankejian@huawei.com>