{topost} IB/hns: Implement the add_gid/del_gid and optimize the GIDs management
IB core has implemented the calculation of GIDs and the management of GID
tables, and it is now responsible to supply query function for GIDs. So
the calculation of GIDs and the management of GID tables in the RoCE
driver is redundant.

The patch is to implement the add_gid/del_gid to set the GIDs in the RoCE
driver, remove the redundant calculation and management of GIDs in the
notifier call of the net device and the inet, and update the query_gid.

Signed-off-by: Shaobo Xu <xushaobo2@huawei.com>
Reviewed-by: Wei Hu (Xavier) <xavier.huwei@huawei.com>