{topost} hisi_sas: use atomic64_t for hisi_sas_device.running_req
Sometimes the value of hisi_sas_device.running_req
may go negative unless we have the check for
running_req >= 0 before trying to decrement.

This is because using running_req is not thread-safe. The
value may actually go negative without the above check.

As such the value for running_req may be incorrect, so
use atomic64_t instead.

Signed-off-by: John Garry <john.garry@huawei.com>
Reviewed-by: Xiang Chen <chenxiang66@hisilicon.com>