{topost} hisi_sas: fix NULL dereference in phy_down_v2_hw()
When the refclock rate is incorrect for v2 HW, we seem
to still get a PHY down event without the initial PHY up.

So, in this scenario, the port in phy_down_v2_hw() is NULL. This
causes a crash as we deference elements in port.

To protect against this, only deference port when port
is non-NULL.

Also some correct values are given for IRQ status in PHY
up/down handler.

Signed-off-by: John Garry <john.garry@huawei.com>
Reviewed-by: Xiang Chen <chenxiang66@hisilicon.com>