{topost} hisi_sas: simplify function check_any_wideports_v2_hw
To check whether there is any wideport, we use complex function to
realize it before. Actually there is a register PORT_STATE which can
check all the ports's status.
If the status of a port is no ready, it presents all the phys in the
port are all down. If the status of a port is ready, it presents at
least one phy is up.
So we can check this register to check there is any wideport.

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Reviewed-by: John Garry <john.garry@huawei.com>