{topost} net: hns: add initialization of xps
The problem is that the performance of UDP on 10 threads is lower when we
run it by the tool named netperf compared with the case of 5 threads.

OS malloc resource and deal delivery on a CPU core when sending a packet.
As the default, the Soc send a packet on the only specified ring.
Then OS will deal the interruption when the packet sending was complete,
include free the memory resource, and that will be and which is implemented
by the only CPU core. This core is unlikely the same with the one sending
packet, so the change of context will cause the fall of UDP performance
when the threads increasing from 5 to 10.

So we use XPS to make sure the core to send packet is the same with the one
to free the memory resource.
Then the UDP performance increases when the threads changed from 5 to 10.
And the performance of UDP increases to 9.6Gb/s on 10 threads,
compared with just 6.5Gb/s before.

Signed-off-by: lipeng <lipeng321@huawei.com>
Signed-off-by: Weiwei Deng <dengweiwei@huawei.com>
Reviewed-by: Yisen Zhuang <yisen.zhuang@huawei.com>