{temphack} libsas: fix the bug of losing disks in hot plug test
When one disk is unpluged, libsas will unregister it and setup
an delay work DISCE_DESTRUCT to destroy the device at last.
If we plug the disk before that delay work finshed, the disk
will be regarded as an new device. At this moment, libsas find
the device is still in domain , then disable its ex phy.

We shouldn't think disk is in domain if the device state is
SAS_DEV_DESTROY or SAS_DEV_GONE.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Reviewed-by: Xiang Chen <chenxiang66@hisilicon.com>