{fromlist v4} firmware: arm_sdei: Discover SDEI support via ACPI
SDEI defines a new ACPI table to indicate the presence of the interface.
The conduit is discovered in the same way as PSCI.

For ACPI we need to create the platform device ourselves as SDEI doesn't
have an entry in the DSDT.

The SDEI platform device should be created after ACPI has been initialised
so that we can parse the table, but before GHES devices are created, which
may register SDE events if they use SDEI as their notification type.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>