{topost} hisi_sas: fix LLDD init problem
There is a potential issue: we initialise PHY before sas_register_ha which we
initialise libsas layer. Though we use 1s timer to delay initialisation of PHY,
there is still a probability that sas_register_ha is not finished after 1s timer
out which will cause OOPS.
To avoid this issue, we initial PHY after sas_register_ha, and also it is not
necessary to use 1s timer.

 Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
 Reviewed-by: John Garry <john.garry@huawei.com>