extio: move the device-relevant data field to strurct extio_ops
There is one field named 'void *devpara' in struct extio_node. It points to
the indirect-IO bus private driver data. For our LPC, it points to struct
hisilpc_dev.

As this field has strong dependency with the I/O hooks defined in struct
extio_ops, it is better to move 'void *devpara' from struct extio_node to
extio_ops. For example, we don't adopt indirect-IO mechanism in our
alternative LPC solution which implement LPC driver as an external module,
then struct extio_node is not needed as it is specific to indirect-IO. In this
case, '*devpara' also need to move to other structure.

I want to keep the driver consistent between the alternative LPC solution and
the community LPC solution. And this modification also makes sense for
lpc-uart earlycon, as the lpc-uart earlycon should call the indirect-IO I/O
accessors after LPC host was probing. With this patch, the earlycon I/O can
use the varaiable hisi_lpc_ops directly to know the struct hisilpc_dev.

Signed-off-by: zhichang.yuan <yuanzhichang@hisilicon.com>