{fromlist v6} drivers: perf: hisi: use poll method to avoid L3C counter overflow
Add hrtimer support which use poll method to avoid counter overflow
when overflow IRQ is not supported in hardware.
The L3 cache PMU use N-N SPI interrupt which has no support in kernel
mainline. So use hrtimer to poll and update event counter to avoid
overflow condition for L3 cache PMU.
An interval of 10 seconds is used for the hrtimer.

Signed-off-by: Dikshit N <dikshit.n@huawei.com>
Signed-off-by: Anurup M <anurup.m@huawei.com>