{topost} ACPI: IORT: SMMUv3: Add support for different SMMU models
After a discussion with Charles with the support for erratums
for SMMU in OEM vendors's design, there is a solution which
we can add different SMMU models for OEM vendors, then use
this number to specify the special handling of the driver.

This patch trys to demo the code and add one example which
Hisilicon SMMUv3 has a erratum which don't support prefetch cmd,
it's suppoerted now with device tree, so use this solution in
ACPI to address the workaround.

Note, the models number will updated when the IORT spec is renewed.

Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>