{fromlist v2} virt: irqbypass: Add a type field to the irqbypass producer
On ARM, we are about to introduce a vfio-platform producer. The
associated architecture specific consumer only is compatible with
this producer and does not work with the dummy vfio-pci msi producer.

Adding a type to the producer allows to easily discriminate between
producers instead of complexifying things by analyzing the interrupt
type.

Signed-off-by: Eric Auger <eric.auger@redhat.com>