{fromlist V3} KVM: arm: Select ARM_GIC_V3 and ARM_GIC_V3_ITS
The GICv4 support introduces a hard dependency between the KVM
core and the ITS infrastructure. arm64 already selects it at
the architecture level, but 32bit doesn't. In order to avoid
littering the kernel with #ifdefs, let's just select the whole
of the GICv3 suport code.

You know you want it.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>