{temphack} scsi: hisi_sas: increase timer expire of internal abort task
In fact, we had increased the time expire from 110ms to 1100ms in
the previous patch "{temphack} hisi_sas: modify expire
of internal abort slot's timer".
But the following issue still can happen: the CQ of internal abort task
is returned after it's timer timeout.

This may cause kernel oops and panic. Because after internal abort task
timeout, we may free the device(Please see the function
hisi_sas_dev_gone()). If this device's residual CQs return at that time,
our driver must be crushed.

And consider the time it may be delayed in tasklet(or softirq).
We increase the timer expire from 1100ms to 20s.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Acked-by: John Garry <john.garry@huawei.com>